modify: separate files
- cpp 파일과 헤더 파일 분리
